python-services/requirements.txt
# Core Financial Libraries
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.2
transformers==4.36.2

# Financial Analysis
QuantLib==1.32
zipline-reloaded==3.0.4
backtrader==1.9.78.123
ta-lib==0.4.28
pyfolio==0.9.2
empyrical==0.5.5
alphalens==0.4.2

# Risk Management
riskfolio-lib==4.3.0
PyPortfolioOpt==1.5.5
cvxpy==1.4.1
riskparityportfolio==0.4.0

# Market Data
alpha-vantage==2.3.1
polygon-api-client==1.12.5
quandl==3.7.0
fredapi==0.5.1
ccxt==4.1.77

# Technical Analysis
talib==0.4.28
finta==1.3
stockstats==0.6.2
pandas-ta==0.3.14b0

# Machine Learning for Finance
xgboost==2.0.3
lightgbm==4.1.0
catboost==1.2.2
optuna==3.5.0
shap==0.44.0

# Time Series
statsmodels==0.14.1
arch==6.2.0
pmdarima==2.0.4
prophet==1.1.5

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
celery==5.3.4
redis==5.0.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
pymongo==4.6.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
websockets==12.0
schedule==1.2.1
python-dateutil==2.8.2
pytz==2023.3

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.39.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
